The AnyBuild cache is significantly smaller that the cache in CloudBuild and in AnyBuild it is desirable to use small (less then 1 Gb) CAS for central storage (given that the checkpoint there is ~200Mb).

+++ Public/Src/Cache/ContentStore/Distributed/NuCache/LocalLocationStoreConfiguration.cs
@@ -423,7 +423,7 @@ namespace BuildXL.Cache.ContentStore.Distributed
         /// <summary>
         /// The maximum number of gigabytes to retain in CAS
         /// </summary>
-        public int MaxRetentionGb { get; set; } = 20;
+        public double MaxRetentionGb { get; set; } = 20;
 
         /// <summary>
         /// Defines the target maximum number of simultaneous copies
